% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_printEquation.R
\name{knit_print.nlmixr2FitCore}
\alias{knit_print.nlmixr2FitCore}
\alias{knit_print.rxUi}
\title{Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.}
\usage{
\method{knit_print}{nlmixr2FitCore}(x, ..., output = "equations")

\method{knit_print}{rxUi}(x, ...)
}
\arguments{
\item{x}{The model to extract equations from}

\item{...}{Ignored}

\item{output}{The type of output to request (currently, just "equations")}
}
\value{
A \code{\link[knitr:asis_output]{knitr::asis_output()}} object containing the 'LaTeX' equation
block, ready to be printed as-is inside a 'knitr'/'rmarkdown' document.
}
\description{
Extract the equations from an nlmixr2/rxode2 model to produce a 'LaTeX'
equation.
}
\examples{
mod <- function() {
  ini({
    lka <- 0.45
    lcl <- 1
    lvc <- 3.45
    propSd <- 0.5
  })
  model({
    ka <- exp(lka)
    cl <- exp(lcl)
    vc <- exp(lvc)
    cp <- linCmt()
    cp ~ prop(propSd)
  })
}
ui <- rxode2::rxode(mod)
knit_print(ui)
}
